clean(core): Move ThingMetadataProvider into new thing.metadata sub-package
