+++ java/dev/enola/cli/CommandWithModel.java
@@ -34,10 +34,10 @@ import dev.enola.core.type.TypeRepositoryBuilder;
 import dev.enola.data.ProviderFromIRI;
 import dev.enola.data.Repository;
 import dev.enola.datatype.DatatypeRepository;
-import dev.enola.thing.ThingMetadataProvider;
 import dev.enola.thing.io.Loader;
 import dev.enola.thing.io.UriIntoThingConverters;
 import dev.enola.thing.message.ThingProviderAdapter;
+import dev.enola.thing.metadata.ThingMetadataProvider;
 import dev.enola.thing.proto.Thing;
 import dev.enola.thing.repo.ThingMemoryRepositoryROBuilder;
 import dev.enola.thing.template.TemplateThingRepository;
