+++ java/dev/enola/thing/gen/markdown/MarkdownSiteGeneratorTest.java
@@ -37,10 +37,10 @@ import dev.enola.datatype.Datatypes;
 import dev.enola.model.enola.files.FileThingConverter;
 import dev.enola.rdf.RdfReaderConverter;
 import dev.enola.rdf.RdfThingConverter;
-import dev.enola.thing.*;
 import dev.enola.thing.impl.ImmutableThing;
 import dev.enola.thing.message.JavaThingToProtoThingConverter;
 import dev.enola.thing.message.ProtoThingIntoJavaThingBuilderConverter;
+import dev.enola.thing.metadata.ThingMetadataProvider;
 import dev.enola.thing.proto.Thing;
 import dev.enola.thing.repo.ThingMemoryRepositoryROBuilder;
 import dev.enola.thing.repo.ThingProvider;
