+++ java/dev/enola/thing/metadata/ThingMetadataProvider.java
@@ -15,13 +15,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.enola.thing;
+package dev.enola.thing.metadata;
 
 import dev.enola.common.io.iri.IRIs;
 import dev.enola.common.io.iri.URIs;
 import dev.enola.common.io.iri.namespace.NamespaceConverter;
 import dev.enola.common.io.metadata.Metadata;
 import dev.enola.common.io.metadata.MetadataProvider;
+import dev.enola.thing.KIRI;
+import dev.enola.thing.Thing;
 import dev.enola.thing.proto.Things;
 import dev.enola.thing.repo.ThingProvider;
 import dev.enola.thing.template.Templates;
